+++ test/IdentityServer4.Contrib.AspNetCore.Testing.Tests/IdentityServerHostProxyTests.cs
@@ -1,5 +1,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Autofac;
+using Autofac.Extensions.DependencyInjection;
 using IdentityServer4.Contrib.AspNetCore.Testing.Builder;
 using IdentityServer4.Contrib.AspNetCore.Testing.Configuration;
 using IdentityServer4.Contrib.AspNetCore.Testing.Services;
@@ -28,7 +30,8 @@ namespace IdentityServer4.Contrib.AspNetCore.Testing.Tests
                 AccessTokenLifetime = 7200
             };
 
-            var hostBuilder = new IdentityServerTestHostBuilder()
+            var hostBuilder = new IdentityServerTestHostCustomContainerBuilder<ContainerBuilder>()
+                .UseServiceProviderFactory(new AutofacServiceProviderFactory(), ContainerBuilderConfiguration.ConfigureContainer)
                 .AddClients(client)
                 .AddApiResources(new ApiResource("api1", "api1name"))
                 .AddApiScopes(new ApiScope("api1"))