+++ test/IdentityServer4.Contrib.AspNetCore.Testing.Tests/Shared/ContainerBuilderConfiguration.cs
@@ -1,0 +1,10 @@
+using Autofac;
+
+namespace IdentityServer4.Contrib.AspNetCore.Testing.Tests
+{
+    public static class ContainerBuilderConfiguration
+    {
+        public static void ConfigureContainer(ContainerBuilder containerBuilder) =>
+            containerBuilder.RegisterType<Dependency>().AsSelf();
+    }
+}