+++ test/IdentityServer4.Contrib.AspNetCore.Testing.Tests/Shared/Dependency.cs
@@ -1,0 +1,6 @@
+namespace IdentityServer4.Contrib.AspNetCore.Testing.Tests
+{
+    public class Dependency
+    {
+    }
+}